Add demo simulating a mail spam detection service
This demo simulates a service that checks if a mail is spam without the
server seeing the mail or the result. If a mail contains "evil.url", it
will be classified as spam.

Add test benches, an unit test and update the README.

Reference: https://vitalik.ca/general/2020/07/20/homomorphic.html