Generate .entry file and optimized IR in Bazel
Use //xls/contrib/xlscc:get_top_func_from_proto to generate .entry files
containing the XLScc entry point for the fhe library.  This information is then
passed to the optimizer and booleanifier.  The booleanifier generates an
%.opt.ir from the %.ir file that XLScc emits, while the booleanifier generates
an %.opt.bool.ir file from the %.opt.ir one created by the optimizer.

This CL is a step towards moving the work to generate IR out of the
fhe_transpiler and into the build system.

PiperOrigin-RevId: 416215725