Introduce the "yosys_plaintext" transpiler as a stub.
With this CL, you can declare transpiler = "yosys_plaintext" in fhe_cc_library, and then
use the %.transpiled_files rule on the target.  Doing this generates Verilog
netlists and a Yosys script, but otherwise pretends to be bool.  The next CL will
modify the transpiler to interpret the netlist.

Adds a cell-definitions library in Liberty format describing the TFHE
primitives.

Also adds a buld target for Yosys and Berkeley ABC (which is used by Yosys).

The change is a no-op for the rest of the FHE codebase.

PiperOrigin-RevId: 416551789