The struct transpiler can now emit struct encodings in source-declaration order.
XLScc reverses the order of struct fields when emitting IR.  When used with FHE
backends that work off of the IR, this does not cause problems. However,
Verilog netlists expect their inputs in the source-language declaration order.

The underlying reason for this order is tuples are flattened to bit vectors in
XLS when emitting verilog, and the ordering of the flattening matches that of
SystemVerilog packed structs where the last field appears in the LSB of the
flattened vector (lowest bit indices). This is also the ordering you get when
you flatten a tuple using the verilog concat operation. Eg:

	a = {tuple_elem0, tuple_elem1, tuple_elem2};

Unfortunately this doesn't match up with struct layout in the C world where the
first field appears at the lowest address.

Add a command-line switch to control the order of fields in the encoded structs.

PiperOrigin-RevId: 416626834