Emit *Ref classes for user-defined data types
For a given UDT named e.g. Polygon, in addition to generating the generic and
scheme-specific classes, we also now emit {Encoded,Palisade,Tfhe}PolygonRef
ones.  The *Ref classes are intended as non-owning references to their
corresponding main types.  This will be used later when converting the
transpiler to emit type-safe function signatures.

PiperOrigin-RevId: 427883483