common_transpiler: add code to generate typed overloads
Add a function to generate type-safe overloads for a given FHE function.  For
example, if you have

	void fibonacci_sequence(int n, int output[5]);

The generated FHE method signature will be (using the bool backend as an
example):

	absl::Status fibonacci_sequence(absl::Span<bool> n, absl::Span<bool> output);

Add a function (not called yet) that will also generate the following type-safe
function:

	absl::Status fibonacci_sequence(const EncodedValueRef<int32_t> n, EncodedArrayRef<int32_t> output) {
		return fibonacci_sequence(n.get(), output.get());
	}

You would be able to invoke the function by passing an EncodedInt as the first
parameters, and an EncodedArray<int32_t> as the second parameter.  The *Ref
conversion will happen implicitly.

This will be used later when the transpiler is converted to emit type-safe
signatures.  (As part of the subsequent CL, the signature of the first function
will take absl::Span<const bool>, rather than absl::Span<bool> as its first
parameter.)

PiperOrigin-RevId: 427920140