Split `transpiler_type` into `optimizer` and `backend`
These are logically equivalent, but avoid the combinatorial explosion we were starting to have in our interface - and make the choice of the XLS-based "Booleanifying optimizer" vs. Yosys more obvious.

We might want to follow up by splitting the "interpreted_" prefix out of the backend parameter, turning it into a boolean parameter of its own. But that's a project for another day.

PiperOrigin-RevId: 428050106