Implement assigning via references
It's now possible to assign a value to a class via a reference to it.  For
example, if you have an instance of Tfhe<Person> and an instance of
TfheRef<Person> pointing to it, you can alter the value of the Tfhe<Person>
instance by altering the value of the reference to it.

Add some more tests.

Introduce an example and a test of simple-image manipulation: sharpening and
Gaussian blur.

PiperOrigin-RevId: 456180997