Remove the `--run_only_passes` and replace all uses with appropriate `--passes` flags
In almost all cases what one wants when using the `--run_only_passes` flag was to run a single pass. Using the `--passes` flag is much clearer. In a small number of cases an actual pipeline is required and in that case the explicit `--passes` pipeline is much nicer and easier to reason about since it doesn't have the implicit embeeding of the standard pipeline.

PiperOrigin-RevId: 597034334